+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Disable clearing the selection while in inspect mode
 - Switch to orthographic projection when maxing out focal length slider, set 18 mm minimum focal length
 - Add button for top-down camera view
+- Make viewport center coordinates editable from the status bar
+
 ## 0.7.0 - 2026-05-12
 
 - Fix issue with colormap clamping past max value